tests: remove schemachange/database-version-upgrade
According to the TODO, this test could have been removed a while back
already.

Release note: None
