+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -113,7 +113,6 @@ go_library(
         "ruby_pg_blocklist.go",
         "rust_postgres.go",
         "rust_postgres_blocklist.go",
-        "schema_change_database_version_upgrade.go",
         "schemachange.go",
         "schemachange_random_load.go",
         "scrub.go",
@@ -161,7 +160,6 @@ go_library(
         "//pkg/cloud",
         "//pkg/cloud/amazon",
         "//pkg/cloud/gcp",
-        "//pkg/clusterversion",
         "//pkg/cmd/cmpconn",
         "//pkg/cmd/roachtest/cluster",
         "//pkg/cmd/roachtest/clusterstats",
